+++ tests/test-site/scripts/build.cjs
@@ -1,6 +1,4 @@
-const { resolve } = require('path');
-const { interpolate, PRPLCachePartitionKey } = require('@prpl/core');
-const { createCachePartition } = require('@prpl/plugin-cache');
+const { interpolate } = require('@prpl/core');
 const { resolveHTMLImports } = require('@prpl/plugin-html-imports');
 const { resolveCSSImports } = require('@prpl/plugin-css-imports');
 const { highlightCode } = require('@prpl/plugin-code-highlight');
@@ -17,24 +15,11 @@ const options = {
 async function build() {
   await interpolate({ options });
 
-  // Pre-define dist partition and use for subsequent plugins
-  await createCachePartition({
-    entityPath: resolve('dist'),
-    partitionKey: PRPLCachePartitionKey.dist,
-    readFileRegExp: new RegExp(`.html|.css`)
-  });
+  await resolveHTMLImports();
 
-  await resolveHTMLImports({
-    cachePartitionKey: PRPLCachePartitionKey.dist
-  });
+  await resolveCSSImports();
 
-  await resolveCSSImports({
-    cachePartitionKey: PRPLCachePartitionKey.dist
-  });
-
-  await highlightCode({
-    cachePartitionKey: PRPLCachePartitionKey.dist
-  });
+  await highlightCode();
 
   const origin = 'http://localhost:8000';
 
@@ -47,8 +32,7 @@ async function build() {
 
   await generateSitemap({
     origin,
-    ignoreDirRegex: new RegExp('dist/fragments'),
-    cachePartitionKey: PRPLCachePartitionKey.dist
+    ignoreDirRegex: new RegExp('dist/fragments')
   });
 }